put websocket upgrade stuff in own file
